Refactor, improve types
