+++ src/webview/WebviewProvider/index.ts
@@ -2,13 +2,7 @@ import * as vscode from "vscode";
 import * as fs from "fs";
 import * as path from "path";
 
-import {
-  buildList,
-  buildTree,
-  fetchPlatformData,
-  getGitHubUrl,
-  fetchHistory
-} from "./lib";
+import { buildList, buildTree, fetchPlatformData, fetchHistory } from "./lib";
 import { AuthProvider, getAccessToken } from "../../auth";
 import { FileNode } from "./lib/workspace/types";
 
@@ -72,8 +66,6 @@ class WebviewProvider implements vscode.WebviewViewProvider {
 
   public async initViewData(refresh?: boolean) {
     const { viewID, _context, _currentView: view } = this;
-    const gitUrl = await getGitHubUrl();
-    console.log("🟢 gitUrl", gitUrl);
     if (!view) return;
     if (viewID === "userInfo") {
       const accessToken = await getAccessToken(_context);
