+++ src/webview/WebviewProvider/lib/index.ts
@@ -4,4 +4,3 @@ export * from "./platform/utils";
 
 export { default as buildList } from "./workspace/buildList";
 export { default as buildTree } from "./workspace/buildTree";
-export { default as getGitHubUrl } from "./workspace/getGitInfo";
