+++ src/webview/WebviewProvider/lib/platform/utils/index.ts
@@ -1,6 +1,6 @@
-export { default as parseResponse } from "./parseResponse";
 export { default as fetchUserInfo } from "./fetchUserInfo";
 export { default as fetchWorkspaces } from "./fetchWorkspaces";
 export { default as fetchComputeEnvs } from "./fetchComputeEnvs";
 export { default as addPipeline } from "./addPipeline";
 export { default as fetchHistory } from "./fetchHistory";
+export { default as getRepoInfo } from "./getRepoInfo";
